Isolate editor
Isolate editor support

Add types and cleanup

with arc

Fix typescript issue

update git ignore

add alias and build step

try

cleanup

Update licenses

Update snapshots
