+++ .eslintrc.json
@@ -24,7 +24,10 @@
         "patterns": [
           "*/neo4j-arc/*",
           "neo4j-arc/common/*",
-          "neo4j-arc/graph-visualization/*"
+          "neo4j-arc/graph-visualization/*",
+          "neo4j-arc/cypher-language-support/*",
+          "^cypher-editor-support$",
+          "^monaco-editor$"
         ]
       }
     ]
@@ -91,8 +94,8 @@
         "no-restricted-imports": [
           "error",
           {
-            "paths": ["redux", "react-redux"],
-            "patterns": ["browser*", "shared*", "services*"]
+            "paths": ["redux", "react-redux", "react-suber"],
+            "patterns": ["browser/*", "shared/*", "services/*"]
           }
         ]
       }
