+++ src/neo4j-arc/cypher-language-support/language/cypherBaseFunctions.ts
