+++ src/neo4j-arc/cypher-language-support/monaco-editor/Monaco.test.tsx
@@ -20,7 +20,7 @@
 import { fireEvent, render } from '@testing-library/react'
 import React from 'react'
 
-import Monaco from './Monaco'
+import { Monaco } from './Monaco'
 
 const noOp = () => undefined
 
@@ -36,9 +36,8 @@ describe('Monaco', () => {
         onExecute={noOp}
         isFullscreen={false}
         toggleFullscreen={noOp}
-        bus={{ self: noOp } as any}
-        params={{}}
         id="id"
+        sendCypherQuery={(() => {}) as any}
       />
     )
