+++ src/neo4j-arc/index.ts
@@ -1,2 +1,3 @@
 export * from './common'
 export * from './graph-visualization'
+export * from './cypher-language-support'
