+++ src/shared/modules/editor/editorDuck.test.ts
@@ -27,7 +27,6 @@ import { COMMAND_QUEUED, executeCommand } from '../commands/commandsDuck'
 import {
   NOT_SUPPORTED_URL_PARAM_COMMAND,
   SET_CONTENT,
-  getText,
   populateEditorFromUrlEpic
 } from './editorDuck'
 
@@ -190,29 +189,5 @@ describe('editorDuck Epics', () => {
 
     // When
     store.dispatch(action)
-  })
-})
-
-describe('getting expected text from cypher-editor-support', () => {
-  test('item with procedure type strips surrounding backticks', () => {
-    const item: EditorSupportCompletionItem = {
-      type: 'procedure',
-      view: '',
-      content: '`apoc.coll.avg`',
-      postfix: null
-    }
-
-    expect(getText(item)).toEqual('apoc.coll.avg')
-  })
-
-  test('item with non procedure or function type retains backticks', () => {
-    const item: EditorSupportCompletionItem = {
-      type: 'label',
-      view: '',
-      content: '`a label name wrapped in backticks`',
-      postfix: null
-    }
-
-    expect(getText(item)).toEqual(item.content)
   })
 })
