+++ tsconfig.json
@@ -31,8 +31,12 @@
       "project-root/*": ["../*"],
       "neo4j-arc/graph-visualization": ["neo4j-arc/graph-visualization"],
       "neo4j-arc/common": ["neo4j-arc/common"],
+      "neo4j-arc/cypher-language-support": [
+        "neo4j-arc/cypher-language-support"
+      ],
       "icons/*": ["browser/icons/*"]
     }
   },
-  "include": ["./src/**/*"]
+  "include": ["./src/**/*"],
+  "exclude": ["./src/neo4j-arc/dist*"]
 }
